+++ addons/l10n_fr_account_ubl_cii/models/__init__.py
@@ -1,0 +1,5 @@
+from . import account_edi_xml_cii_facturx
+from . import account_edi_xml_ubl_21_fr
+from . import account_move
+from . import res_company
+from . import res_partner
