+++ addons/l10n_fr_account_ubl_cii/tests/__init__.py
@@ -1,0 +1,3 @@
+from . import common
+from . import test_partner
+from . import test_xml_ubl_21_fr
