+++ addons/l10n_fr_account_ubl_cii/tests/test_files/export/ubl_21_fr/invoice/fr/ubl_21_fr_out_invoice.xml
