+++ addons/l10n_fr_pdp/models/__init__.py
@@ -1,6 +1,4 @@
 from . import account_edi_proxy_user
-from . import account_edi_xml_cii_facturx
-from . import account_edi_xml_ubl_21_fr
 from . import account_journal
 from . import account_move
 from . import account_move_send
