+++ addons/l10n_fr_pdp/models/account_move.py
@@ -4,10 +4,9 @@ from datetime import datetime
 
 from odoo import api, fields, models
 from odoo.exceptions import UserError, ValidationError
-from odoo.tools import frozendict, html2plaintext
+from odoo.tools import frozendict
 
 from odoo.addons.l10n_fr_pdp.models.account_edi_proxy_user import STATUS_TO_PROCESS_CONDITION_CODE_PDP
-from odoo.addons.l10n_fr_pdp.models.account_edi_xml_ubl_21_fr import PDP_CUSTOMIZATION_ID
 from odoo.addons.l10n_fr_pdp.models.account_peppol_response import NEW_STATUSES
 from odoo.addons.l10n_fr_pdp.models.pdp_flow import FLOW_OPEN_STATES_SELECTION, FLOW_SENT_STATES, FLOW_SENT_STATES_SELECTION
 from odoo.addons.l10n_fr_pdp.utils import drom_com_territories
@@ -236,31 +235,6 @@ class AccountMove(models.Model):
         if {'refused', 'RE'} & states:
             return 'error'
         return 'sent'
-
-    def _l10n_fr_pdp_get_default_notes(self):
-        self.ensure_one()
-        # Mandatory / default notes for French e-invoicing [BR-FR-05]
-        # Only add them when using PDP
-        if self.company_id._get_peppol_proxy_type() != 'pdp':
-            return {}
-        payment_term = self.invoice_payment_term_id
-        return {
-            'PMT': self.env._("In the event of late payment, a flat-rate fee of €40 for collection costs will be charged (Articles L.441-10 and D.441-5 of the Code de commerce)."),
-            'PMD': self.env._("Late payment penalties at an annual rate of 10% are applied if the payment is made after the due date."),
-            'AAB': html2plaintext(payment_term.note) if payment_term.early_discount else self.env._("No discount for early payment."),
-        }
-
-    @api.model
-    def _get_ubl_cii_builder_from_xml_tree(self, tree):
-        # Extends account_edi_ubl_cii
-        customization_id = tree.find('{*}CustomizationID')
-        # Note: The CustomizationID alone is not enough because e.g. SuperPDP just sends `urn:cen.eu:en16931:2017`
-        #       but still expects the full French validation.
-        if customization_id is not None and customization_id.text == PDP_CUSTOMIZATION_ID:
-            receiver_endpoint_node = tree.find('./{*}AccountingCustomerParty/{*}Party/{*}EndpointID')
-            if receiver_endpoint_node is not None and receiver_endpoint_node.get('schemeID') == '0225':
-                return self.env['account.edi.xml.ubl_21_fr']
-        return super()._get_ubl_cii_builder_from_xml_tree(tree)
 
     def action_pdp_open_response_wizard(self, **wizard_kwargs):
         if not (pdp_moves := self.filtered('pdp_can_send_response')):
