+++ addons/l10n_fr_pdp/models/account_move_send.py
@@ -8,7 +8,7 @@ class AccountMoveSend(models.AbstractModel):
     def _get_default_sending_method(self, move) -> str:
         # EXTENDS 'account_peppol' to not select Peppol / PDP for B2C invoices
         partner = move.commercial_partner_id.with_company(move.company_id)
-        if move.company_id._get_peppol_proxy_type() != 'pdp' or not partner._l10n_fr_pdp_is_b2c():
+        if move.company_id._get_peppol_proxy_type() != 'pdp' or not partner._l10n_fr_is_b2c():
             return super()._get_default_sending_method(move)
         return 'email'
