+++ addons/l10n_fr_pdp/tests/__init__.py
@@ -3,4 +3,3 @@ from . import test_flow_lifecycle
 from . import test_messages
 from . import test_partner
 from . import test_user
-from . import test_xml_ubl_21_fr
